requirements.txt
asgiref==3.11.1
Django==6.0.3
sqlparse==0.5.5

# Supabase PostgreSQL + Storage (S3 API)
dj-database-url>=2.2,<3
psycopg[binary]>=3.1,<4
django-storages[s3]>=1.14,<2
boto3>=1.34
python-dotenv>=1.0

# Produção (ex.: Render)
gunicorn>=22.0
